Rocket/requirements.txt
aiosqlite==0.19.0
LabJackPython==2.1.0
numpy==1.25.2
pyserial==3.5
pyserial-asyncio==0.6
Jetson.GPIO==2.1.4